:hammer: refactor: WinningStatistics 패키지 변경 & 개행 수정
